requirements.txt
# Backend runtime only (used by Docker / Azure)
fastapi>=0.115.0,<1
uvicorn[standard]>=0.32.0,<1
pydantic>=2.9.0,<3
pydantic-settings>=2.6.0,<3
openai>=1.55.0,<2
httpx>=0.27.0,<1
huggingface_hub>=0.27.0,<1
sentence-transformers>=3.0.0,<4
numpy>=1.26.0,<3
pillow>=10.4.0,<12
